+++ src/finalizer.rs
@@ -14,7 +14,7 @@ use miniscript::{bitcoin, plan::Plan, psbt::PsbtInputSatisfier};
 ///
 /// # Usage
 ///
-/// A [`Finalizer`] is typically obtained alongside a PSBT from [`TxTemplate::create_psbt`].
+/// A [`Finalizer`] is typically obtained alongside a PSBT from [`TxTemplate::build_psbt`].
 /// It can also be constructed directly from a list of `(outpoint, plan)` pairs via
 /// [`Finalizer::new`]. Use [`finalize_input`] to finalize a single input, or [`finalize`] to
 /// finalize every input. Upon finalizing the PSBT, the [`Finalizer`] also clears metadata from
@@ -23,12 +23,12 @@ use miniscript::{bitcoin, plan::Plan, psbt::PsbtInputSatisfier};
 /// # Example
 ///
 /// ```rust,no_run
-/// # use bdk_tx::PsbtBuildParams;
+/// # use bdk_tx::BuildPsbtParams;
 /// # let secp = bitcoin::secp256k1::Secp256k1::new();
 /// # let keymap = std::collections::BTreeMap::new();
 /// # let template: bdk_tx::TxTemplate = unimplemented!();
 /// // Create PSBT from a selection of inputs and outputs.
-/// let (mut psbt, finalizer) = template.create_psbt(PsbtBuildParams::default())?;
+/// let (mut psbt, finalizer) = template.build_psbt(BuildPsbtParams::default())?;
 ///
 /// // Sign the PSBT using your preferred method.
 /// let signer = bdk_tx::Signer(keymap);
@@ -45,7 +45,7 @@ use miniscript::{bitcoin, plan::Plan, psbt::PsbtInputSatisfier};
 ///
 /// [BIP174]: <https://github.com/bitcoin/bips/blob/master/bip-0174.mediawiki#input-finalizer>
 /// [`TxTemplate`]: crate::TxTemplate
-/// [`TxTemplate::create_psbt`]: crate::TxTemplate::create_psbt
+/// [`TxTemplate::build_psbt`]: crate::TxTemplate::build_psbt
 /// [`Plan`]: miniscript::plan::Plan
 /// [`Transaction`]: bitcoin::Transaction
 /// [`finalize_input`]: Finalizer::finalize_input
@@ -163,7 +163,7 @@ impl FinalizeMap {
 #[cfg_attr(coverage_nightly, coverage(off))]
 #[cfg(test)]
 mod tests {
-    use crate::{Finalizer, Output, PsbtBuildParams, Signer, TxTemplate};
+    use crate::{BuildPsbtParams, Finalizer, Output, Signer, TxTemplate};
     use bitcoin::secp256k1::Secp256k1;
     use bitcoin::{absolute, transaction, Amount, ScriptBuf, TxIn, TxOut};
     use miniscript::bitcoin;
@@ -217,7 +217,7 @@ mod tests {
         let output = Output::with_script(ScriptBuf::new(), Amount::from_sat(9_000));
         let selection = TxTemplate::from_parts(vec![input], vec![output]);
 
-        let (mut psbt, finalizer) = selection.create_psbt(PsbtBuildParams::default())?;
+        let (mut psbt, finalizer) = selection.build_psbt(BuildPsbtParams::default())?;
 
         let secp = Secp256k1::new();
         let signer = Signer(keymap);
@@ -236,7 +236,7 @@ mod tests {
         let output = Output::with_script(ScriptBuf::new(), Amount::from_sat(9_000));
         let selection = TxTemplate::from_parts(vec![input], vec![output]);
 
-        let (mut psbt, finalizer) = selection.create_psbt(PsbtBuildParams::default())?;
+        let (mut psbt, finalizer) = selection.build_psbt(BuildPsbtParams::default())?;
 
         let secp = Secp256k1::new();
         let signer = Signer(keymap);
@@ -264,7 +264,7 @@ mod tests {
             ],
         );
 
-        let (mut psbt, finalizer) = selection.create_psbt(PsbtBuildParams::default())?;
+        let (mut psbt, finalizer) = selection.build_psbt(BuildPsbtParams::default())?;
 
         assert!(!psbt.outputs[0].tap_key_origins.is_empty());
         assert!(psbt.outputs[0].tap_internal_key.is_some());
@@ -318,7 +318,7 @@ mod tests {
             ],
         );
 
-        let (mut psbt, _) = selection.create_psbt(PsbtBuildParams::default())?;
+        let (mut psbt, _) = selection.build_psbt(BuildPsbtParams::default())?;
 
         let tap_key_origins = psbt.outputs[0].tap_key_origins.clone();
         let tap_internal_key = psbt.outputs[0].tap_internal_key;
@@ -358,7 +358,7 @@ mod tests {
             ],
         );
 
-        let (mut psbt, finalizer) = selection.create_psbt(PsbtBuildParams::default())?;
+        let (mut psbt, finalizer) = selection.build_psbt(BuildPsbtParams::default())?;
 
         let tap_key_origins = psbt.outputs[0].tap_key_origins.clone();
         let tap_internal_key = psbt.outputs[0].tap_internal_key;
@@ -385,7 +385,7 @@ mod tests {
         let output = Output::with_script(ScriptBuf::new(), Amount::from_sat(9_000));
         let selection = TxTemplate::from_parts(vec![input], vec![output]);
 
-        let (mut psbt, finalizer) = selection.create_psbt(PsbtBuildParams::default())?;
+        let (mut psbt, finalizer) = selection.build_psbt(BuildPsbtParams::default())?;
 
         let secp = Secp256k1::new();
         let signer = Signer(keymap);
